메모리 상태값 변경

+++ compute-service/src/main/kotlin/com/ttalkak/compute/compute/adapter/in/socket/request/ComputeStatusRequest.kt
@@ -6,7 +6,7 @@ data class ComputeStatusRequest (
     val userId: Long,
     val computerType: ComputerType,
     val usedCompute: Int,
-    val usedMemory: Int,
+    val usedMemory: Long,
     val usedCPU: Double,
     val deployments: List<DeploymentStatusRequest>
 )
