+++ compute-service/src/main/kotlin/com/ttalkak/compute/compute/application/port/in/AllocateCommand.kt
@@ -2,5 +2,5 @@ package com.ttalkak.compute.compute.application.port.`in`
 
 data class AllocateCommand(
     val computeCount: Int,
-    val useMemory: Int,
+    val useMemory: Double,
 )
