+++ compute-service/src/main/kotlin/com/ttalkak/compute/compute/application/port/in/ConnectCommand.kt
@@ -6,6 +6,6 @@ data class ConnectCommand(
     val userId: Long,
     val computeType: ComputerType,
     val usedCompute: Int,
-    val usedMemory: Int,
+    val usedMemory: Double,
     val usedCPU: Double,
 )
