+++ compute-service/src/main/kotlin/com/ttalkak/compute/compute/application/port/in/StatusCommand.kt
@@ -2,7 +2,7 @@ package com.ttalkak.compute.compute.application.port.`in`
 
 data class StatusCommand(
     val maxCompute: Int,
-    val maxMemory: Int,
+    val maxMemory: Double,
     val maxCPU: Double,
     val availablePortStart: Int,
     val availablePortEnd: Int
